requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
psycopg2-binary==2.9.12
redis==7.4.0
python-dotenv==1.2.2